+++ libs/ngx-contextmenu/.storybook/main.ts
@@ -1,9 +1,11 @@
+import { fileURLToPath } from "node:url";
+import { dirname } from "node:path";
 import type { StorybookConfig } from '@storybook/angular';
 import remarkGfm from 'remark-gfm';
 
 const config: StorybookConfig = {
   framework: {
-    name: '@storybook/angular',
+    name: getAbsolutePath("@storybook/angular"),
     options: {},
   },
 
@@ -11,7 +13,7 @@ const config: StorybookConfig = {
 
   addons: [
     {
-      name: '@storybook/addon-docs',
+      name: getAbsolutePath("@storybook/addon-docs"),
       options: {
         mdxPluginOptions: {
           mdxCompileOptions: {
@@ -33,3 +35,7 @@ const config: StorybookConfig = {
 };
 
 export default config;
+
+function getAbsolutePath(value: string): any {
+  return dirname(fileURLToPath(import.meta.resolve(`${value}/package.json`)));
+}